Fix commit message in non-UTF-8 encoding may be shown corrupted.
Commit messages can have an encoding different from UTF-8. There is a machinery
that auto-detects encoding for the text. But the latter can have incorrect
results. As for commit messages, they have a definite encoding that a raw
commit object keeps explicitly. Used `rugged` gem sets a commit message property to
this encoding. So, there is not need to detect commit message encoding, we just
need to convert a text from this encoding to UTF-8 (that is used for output).

Gitaly since a655968 returns commit message as raw bytes (to be fixed).

This fixes #39819.

This patch needs commit 0fdaacec726e4ab60f9f4df80a4521f8568e4ae7 (empty commit
made in Windows-1251 encoding) in a branch named `commit-encoding-windows-1251`
in `gitlab-test` repository.